+++ CKODemoShop/ClientApp/src/app/components/hypermedia/hypermedia.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { ILinks } from 'src/app/interfaces/links.interface';
-import { OrderService } from 'src/app/services/order.service';
 import { IPayment } from 'src/app/interfaces/payment.interface';
+import { PaymentsService } from 'src/app/services/payments.service';
 
 @Component({
   selector: 'app-hypermedia',
@@ -12,12 +12,12 @@ export class HypermediaComponent implements OnInit {
   links: ILinks;
   actions: Object[];
 
-  constructor(private _orderService: OrderService) { }
+  constructor(private _paymentsService: PaymentsService) { }
 
   ngOnInit() {
     this.links = this.payment._links;
     if (this.links.actions) {
-      this._orderService.getPaymentActions(this.payment.id).subscribe(
+      this._paymentsService.getPaymentActions(this.payment.id).subscribe(
         response => this.actions = response.body
       );
     }
